pyproject.toml
[project]
name = "dagster_tailscale_devices"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "dagster",
    "dagster-cloud",
    "dagster-pandas",
    "psycopg",
]

[project.optional-dependencies]
dev = [
    "dagster-webserver", 
    "pytest",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.dagster]
module_name = "dagster_tailscale_devices.definitions"
project_name = "dagster_tailscale_devices"

[tool.setuptools.packages.find]
exclude=["dagster_tailscale_devices_tests"]
